dealer: Configure db via flags and env using ff pkg
