+++ services/dealer/.gitignore
@@ -1,1 +1,1 @@
-dealer
+./dealer
